various refactoring and cleaned up code
student: Ace
mentor: Robert
