+++ src/main/java/frc/robot/lib/NavX.java
@@ -3,7 +3,7 @@
 // Open Source Software; you can modify and/or share it under the terms of BSD
 // license file in the root directory of this project.
 
-package frc.robot.utils;
+package frc.robot.lib;
 
 import com.kauailabs.navx.frc.AHRS;
