--- src/main/java/frc/robot/subsystems/VisionTracking.java
@@ -1,18 +1,0 @@
-// Copyright (c) 2023 FRC Team 2881 - The Lady Cans
-//
-// Open Source Software; you can modify and/or share it under the terms of BSD
-// license file in the root directory of this project.
-
-package frc.robot.subsystems;
-
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-
-public class VisionTracking extends SubsystemBase {
-  /** Creates a new VisionTracking. */
-  public VisionTracking() {}
-
-  @Override
-  public void periodic() {
-    // This method will be called once per scheduler run
-  }
-}
